+++ rspec-expectations/lib/rspec/matchers/built_in/contain_exactly.rb
@@ -3,10 +3,10 @@
 module RSpec
   module Matchers
     module BuiltIn
-      # rubocop:disable Metrics/ClassLength
       # @api private
       # Provides the implementation for `contain_exactly` and `match_array`.
       # Not intended to be instantiated directly.
+      # rubocop:disable Metrics/ClassLength
       class ContainExactly < BaseMatcher
         # @api private
         # @return [String]
@@ -109,19 +109,8 @@ module RSpec
           array
         end
 
-        if RUBY_VERSION == "1.8.7"
-          # :nocov:
-          def to_a_disallowed?(object)
-            case object
-            when NilClass, String then true
-            else Kernel == RSpec::Support.method_handle_for(object, :to_a).owner
-            end
-          end
-          # :nocov:
-        else
-          def to_a_disallowed?(object)
-            NilClass === object
-          end
+        def to_a_disallowed?(object)
+          NilClass === object
         end
 
         def missing_items