+++ rspec-expectations/lib/rspec/matchers/built_in/has.rb
@@ -71,25 +71,12 @@ module RSpec
           really_responds_to?(predicate)
         end
 
-        # support 1.8.7, evaluate once at load time for performance
-        if String === methods.first
-          # :nocov:
-          def private_predicate?
-            @actual.private_methods.include? predicate.to_s
-          end
+        def private_predicate?
+          @actual.private_methods.include? predicate
+        end
 
-          def methods_include?(method)
-            @actual.methods.include?(method.to_s)
-          end
-          # :nocov:
-        else
-          def private_predicate?
-            @actual.private_methods.include? predicate
-          end
-
-          def methods_include?(method)
-            @actual.methods.include?(method)
-          end
+        def methods_include?(method)
+          @actual.methods.include?(method)
         end
 
         def predicate_result
@@ -109,11 +96,7 @@ module RSpec
         end
 
         def root
-          # On 1.9, there appears to be a bug where String#match can return `false`
-          # rather than the match data object. Changing to Regex#match appears to
-          # work around this bug. For an example of this bug, see:
-          # https://travis-ci.org/rspec/rspec-expectations/jobs/27549635
-          self.class::REGEX.match(@method_name.to_s).captures.first
+          self.class::REGEX.match(@method_name).captures.first
         end
 
         def method_description