+++ rspec-expectations/spec/rspec/expectations/failure_aggregator_spec.rb
@@ -455,8 +455,6 @@ module RSpec::Expectations
         fail { |e| expect(e.message).to include(yield) }
       end
 
-      # rubocop:disable Lint/DuplicateBranch
-
       # Each Ruby version return a different exception complement.
       # This method gets the current version and return the
       # right complement.
@@ -464,28 +462,15 @@ module RSpec::Expectations
         def exception_complement(block_levels)
           ":in 'block (#{block_levels} levels) in <module:Expectations>'"
         end
-      elsif RSpec::Support::Ruby.mri? && RUBY_VERSION > "1.8.7"
+      elsif RSpec::Support::Ruby.mri? || RSpec::Support::Ruby.truffleruby?
         def exception_complement(block_levels)
           ":in `block (#{block_levels} levels) in <module:Expectations>'"
-        end
-      elsif RSpec::Support::Ruby.mri?
-        def exception_complement(block_levels)
-          ""
-        end
-      elsif RSpec::Support::Ruby.truffleruby?
-        def exception_complement(block_levels)
-          ":in `block (#{block_levels} levels) in <module:Expectations>'"
-        end
-      elsif RUBY_VERSION > "2.0.0"
-        def exception_complement(block_levels)
-          ":in `block in Expectations'"
         end
       else
         def exception_complement(block_levels)
-          ":in `Expectations'"
+          ":in `block in Expectations'"
         end
       end
-      # rubocop:enable Lint/DuplicateBranch
     end
   end
 end